custom-attack/requirements.txt
blinker==1.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-nmap==0.7.1
requests==2.32.3
setuptools==71.1.0
six==1.16.0
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.2
wheel==0.43.0
dnspython